Prototype implementation for netlink link status
This watches netlink link status instead of polling probe of link
status. This event is only happen if link status is changed, so
could be efficient than polling.

This patch only gives the routine for netlink, so it does not remove
Probe() yet. Comitter need to consider whether Probe() still useful
or not...